+++ lib/omdb/api.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require 'omdb/api/version'
+require 'omdb/api/error'
 require 'omdb/api/client'